# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Portfolio API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/projects": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Display a listing of projects."
                    }
                }
            },
            "post": {
                "requestBody": [
                    ""
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of projects."
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "parameters": "id",
                "responses": {
                    "200": {
                        "description": "Display a listing of projects."
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}